app/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
openai==1.75.0
anthropic==0.49.0
google-cloud-aiplatform==1.91.0
python-dotenv==1.0.0
google-cloud-firestore==2.19.0
pyyaml==6.0.2